Use built-in Node.js test/assert libraries
This commit replaces the AVA test library with Node.js' built-in test
and assertion libraries. For this library's purposes, the built-in
assertions will do the trick.

This change dramatically reduces the number of dependencies and improves
installation time.

The test file itself was refactored slightly to accommodate the change
and to align constant naming conventions with the various project's
example files (which was helpful for someone like me who isn't
terribly familiar with AJV, etc.).